declaration for circle.py fixed, README.md updated

+++ circle.py
@@ -7,7 +7,7 @@ def area(r):
             Parameters:
                 r (int): radius of the circle
             Returns:
-                (int): area of the circle
+                (float): area of the circle
     '''
 
     return math.pi * r * r
@@ -19,7 +19,7 @@ def perimeter(r):
             Parameters:
                 r (int): radius of the circle
             Returns:
-                (int): perimeter of the circle
+                (float): perimeter of the circle
     '''
 
     return 2 * math.pi * r
